+++ src/Symfony/Bundle/SecurityBundle/EventListener/AclSchemaListener.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Bundle\SecurityBundle\EventListener;
 
+@trigger_error(sprintf('Class "%s" is deprecated since version 3.4 and will be removed in 4.0. Use Symfony\Bundle\AclBundle\EventListener\SetAclCommand instead.', AclSchemaListener::class), E_USER_DEPRECATED);
+
 use Symfony\Component\Security\Acl\Dbal\Schema;
 use Doctrine\ORM\Tools\Event\GenerateSchemaEventArgs;
 
@@ -18,6 +20,8 @@ use Doctrine\ORM\Tools\Event\GenerateSchemaEventArgs;
  * Merges ACL schema into the given schema.
  *
  * @author Johannes M. Schmitt <schmittjoh@gmail.com>
+ *
+ * @deprecated since 3.4, to be removed in 4.0
  */
 class AclSchemaListener
 {